# desktop/src/open-api/model/createRecognitionTaskCommand.ts
/**
 * Receipt Wrangler API.
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ReceiptStatus } from './receiptStatus';


export interface CreateRecognitionTaskCommand { 
    clientRequestId: string;
    fileName: string;
    fileSize: number;
    groupId: number;
    paidByUserId?: number;
    status?: ReceiptStatus;
    categoryIds?: Array<number>;
    tagIds?: Array<number>;
}
export namespace CreateRecognitionTaskCommand {
}